refactor(replica): move resumable reader internal
